adding visualization tool as a separate executable, added dynamically maintained fill factor, added spot growth, refactored to hopefully make parallel but doesn't seem to work right now

+++ star.hpp
@@ -32,7 +32,7 @@ public:
     Profile profile_quiet, profile_active;
     std::vector<double> integrated_ccf;
     std::vector<double> fit_result;
-    std::vector<double> image;
+    std::vector<float> image;
 
 private:
     double equatorial_velocity;
